feat!: mutate single node, check depth when generating a new tree
It looks like the norm is to mutate a single node.

When mutating, if we generate a random tree, instead of using a fixed
max depth (2 was used to keep the new tree shallow), compute the maximum
depth that could be used for current node (by using its depth)

Also I was mixing up the depth with the height. We assume there are
"only" nodes, so the depth is from the root to the current node and the
height is from the current node to the deepest leaf.

To introduce the actual depth, we had to introduce a new attribute: the
parent.